requirements.txt
injector==0.21.0
pytest>=7.4.2
vaultcredentialsmanager>=1.0.10
psycopg2==2.9.9
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-wsgi==0.41b0
honeycomb-opentelemetry==0.2.3b0
jwcrypto==1.5.1
PyJWT
alembic==1.13.1
SQLAlchemy==2.0.31
Flask==3.0.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
flask-restx>=1.3.0
Flask-Injector == 0.15.0


